+++ rpc/backend/utils_test.go
@@ -1,0 +1,54 @@
+package backend
+
+import (
+	"fmt"
+
+	"github.com/tendermint/tendermint/proto/tendermint/crypto"
+)
+
+func mookProofs(num int, withData bool) *crypto.ProofOps {
+	var proofOps *crypto.ProofOps
+	if num > 0 {
+		proofOps = &crypto.ProofOps{
+			Ops: make([]crypto.ProofOp, num),
+		}
+		for i := 0; i < num; i++ {
+			proof := crypto.ProofOp{}
+			if withData {
+				proof.Data = []byte("ops:<type:\"ics23:iavl\" key:\"\\001\\177\\r\\025\\307\\372\\256e\\211fH\\310';m~C\\365\\217\\250B\" data:\"\\022\\314\\021\\n\\025\\001\\177\\r\\025\\307\\372\\256e\\211fH\\310';m~C\\365\\217\\250B\\022\\332\\010\\n\\025\\001\\177\\r\\021X\\232W\\326\\274\\014\\250fl*w\\003\\230k\\235Rq\\022\\233\\001\\n\\036/ethermint.types.v1.EthAccount\\022y\\n3\\n+tcrc10ux3zky62lttcr9gvekz5acrnp4e65n3qw284f\\030\\260\\312\\036 '\\022B0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470\\032\\016\\010\\001\\030\\001 \\001*\\006\\000\\002\\262\\324\\376\\001\\\".\\010\\001\\022\\007\\002\\004\\262\\324\\376\\001 \\032! )a\\251\\226S\\263\\216:\\370\\321l\\2402$\\246\\034\\314\\350\\265\\0170\\3500V\\004\\252\\320IT\\324j\\236\\\",\\010\\001\\022(\\004\\006\\262\\324\\376\\001 \\002\\201%9o\\237\\275\\204\\035f-\\204\\003|\\245\\t\\010\\203M\\233U~5\\303\\\"u\\310\\006\\232(Q\\211 \\\",\\010\\001\\022(\\006\\n\\262\\324\\376\\001 \\304-!\\322/\\241\\314\\025\\340b\\272V\\252\\347\\224\\200\\313\\345\\021(\\270\\202\\3072u\\034\\271Om\\005\\003\\037 \\\",\\010\\001\\022(\\n\\034\\300\\336\\222\\004 \\250\\217\\216_\\227\\274\\271\\336\\210\\343\\277o\\025\\301\\037-\\310\\207d\\363\\361L\\253\\357\\2075Yag8\\335\\242 \\\".\\010\\001\\022\\007\\014H\\300\\336\\222\\004 \\032! \\214?\\016\\2351|\\265Z\\347\\275\\245\\210Gm\\3161\\n\\220\\316\\243:b\\227\\350\\026\\306n\\020\\356n\\243\\300\\\"-\\010\\001\\022)\\016\\226\\001\\300\\336\\222\\004 \\322$OZ\\002\\255$j<\\031\\323\\357\\177\\366\\362%\\275Wk\\001\\355@\\312\\003&^f65I\\251g \\\"/\\010\\001\\022\\010\\022\\334\\002\\310\\230\\322\\005 \\032! G\\001\\373 RM\\212e(\\357\\223W;\\202\\270\\377\\336\\r\\022;J\\034!\\324\\352\\027\\302|\\022d\\366u\\\"/\\010\\001\\022\\010\\024\\306\\004\\310\\230\\322\\005 \\032! \\347\\260\\0323]\\031C\\023\\243\\361\\357\\355'g\\206\\227\\250\\261v\\347\\363\\341\\202h\\021Y\\2032v\\357`\\233\\\"/\\010\\001\\022\\010\\026\\226\\010\\310\\230\\322\\005 \\032! \\017\\342U\\306Z\\245\\nA\\262w\\205R\\320l\\251gw<X\\203x}\\236?4\\210Z\\324d\\342\\236\\310\\\"-\\010\\001\\022)\\030\\364\\024\\310\\230\\322\\005 q\\203V\\221\\0335\\351\\326\\020H\\351Q\\360\\344\\350\\\"\\350G\\037:\\034\\313\\335\\260zZM\\220\\211\\353\\355\\357 \\\"-\\010\\001\\022)\\032\\350&\\212\\321\\322\\005 U\\267\\016\\204\\032\\361+\\271\\\"\\245\\272N\\021\\377\\313%\\375C=\\306\\256\\367\\266\\005\\033r*a\\341\\375\\232\\014 \\\"/\\010\\001\\022\\010\\034\\240B\\212\\321\\322\\005 \\032! \\355!\\316\\022tW\\031c\\260\\263e2\\\"\\345i\\203\\227\\ryP^\\001[\\326\\027\\tuQ\\235\\370\\302v\\\"0\\010\\001\\022\\t\\036\\324\\205\\001\\346\\234\\323\\005 \\032! \\335B\\276M7\\3173\\3035V'\\033\\315\\244\\200\\330 \\237j\\314b\\253\\352\\246\\304^D\\325\\345\\350\\310[\\\".\\010\\001\\022* \\200\\204\\002\\350\\260\\323\\005 \\352\\300\\001*\\345\\245\\225\\253K#//\\244\\256\\253\\327t\\023\\246\\302O\\363c\\212W\\031\\032\\3467l\\214B \\\"0\\010\\001\\022\\t$\\216\\316\\004\\350\\260\\323\\005 \\032! \\3338:Q (\\210u\\332l>\\374\\222\\310d6\\205\\321\\000\\231O\\263\\367\\266n\\372\\311\\034B\\363X\\234\\\".\\010\\001\\022*&\\366\\305\\007\\350\\260\\323\\005 \\354\\373\\303\\317\\271\\\\\\367d\\252\\346D3\\205\\244\\221\\207\\363\\346\\326\\301\\202\\203/X\\321\\251\\022\\310\\277\\241\\0308 \\\"0\\010\\001\\022\\t*\\240\\276\\022\\356\\260\\323\\005 \\032! \\027\\355I\\365Cg\\243\\374VL\\350\\003\\033Rjt\\364\\231\\375\\311\\223\\005V\\225%\\262BOG\\315\\363\\034\\\"1\\010\\001\\022\\n.\\220\\203\\206\\001\\356\\260\\323\\005 \\032! I\\262\\237\\177\\320\\342\\211\\311\\273C\\314C\\350)\\356\\257Z\\266y\\377\\315\\226y\\302\\306E\\266\\021\\257q\\322\\300\\\"/\\010\\001\\022+0\\262\\336\\254\\001\\356\\260\\323\\005 \\207{;n\\317\\337}\\2446\\235\\030dz\\367\\373\\200\\021\\014\\276\\330g\\026\\330V7\\374WM\\221\\3142# \\032\\325\\010\\n\\025\\001\\177\\rp\\000\\330P~\\350(C\\220M\\251[#b\\262\\321\\341\\341\\022\\231\\001\\n\\036/ethermint.types.v1.EthAccount\\022w\\n1\\n+tcrc10uxhqqxc2plws2zrjpx6jkerv2edrc0p9fuumn\\030\\260\\355\\023\\022B0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470\\032\\r\\010\\001\\030\\001 \\001*\\005\\000\\002\\334\\366\\026\\\",\\010\\001\\022(\\002\\004\\262\\324\\376\\001 c\\376\\000\\225\\204\\017t\\320'\\n\\345\\366:?v-\\273\\201\\215\\315%\\rV\\230-\\243+\\277\\315\\036\\272( \\\",\\010\\001\\022(\\004\\006\\262\\324\\376\\001 \\002\\201%9o\\237\\275\\204\\035f-\\204\\003|\\245\\t\\010\\203M\\233U~5\\303\\\"u\\310\\006\\232(Q\\211 \\\",\\010\\001\\022(\\006\\n\\262\\324\\376\\001 \\304-!\\322/\\241\\314\\025\\340b\\272V\\252\\347\\224\\200\\313\\345\\021(\\270\\202\\3072u\\034\\271Om\\005\\003\\037 \\\",\\010\\001\\022(\\n\\034\\300\\336\\222\\004 \\250\\217\\216_\\227\\274\\271\\336\\210\\343\\277o\\025\\301\\037-\\310\\207d\\363\\361L\\253\\357\\2075Yag8\\335\\242 \\\".\\010\\001\\022\\007\\014H\\300\\336\\222\\004 \\032! \\214?\\016\\2351|\\265Z\\347\\275\\245\\210Gm\\3161\\n\\220\\316\\243:b\\227\\350\\026\\306n\\020\\356n\\243\\300\\\"-\\010\\001\\022)\\016\\226\\001\\300\\336\\222\\004 \\322$OZ\\002\\255$j<\\031\\323\\357\\177\\366\\362%\\275Wk\\001\\355@\\312\\003&^f65I\\251g \\\"/\\010\\001\\022\\010\\022\\334\\002\\310\\230\\322\\005 \\032! G\\001\\373 RM\\212e(\\357\\223W;\\202\\270\\377\\336\\r\\022;J\\034!\\324\\352\\027\\302|\\022d\\366u\\\"/\\010\\001\\022\\010\\024\\306\\004\\310\\230\\322\\005 \\032! \\347\\260\\0323]\\031C\\023\\243\\361\\357\\355'g\\206\\227\\250\\261v\\347\\363\\341\\202h\\021Y\\2032v\\357`\\233\\\"/\\010\\001\\022\\010\\026\\226\\010\\310\\230\\322\\005 \\032! \\017\\342U\\306Z\\245\\nA\\262w\\205R\\320l\\251gw<X\\203x}\\236?4\\210Z\\324d\\342\\236\\310\\\"-\\010\\001\\022)\\030\\364\\024\\310\\230\\322\\005 q\\203V\\221\\0335\\351\\326\\020H\\351Q\\360\\344\\350\\\"\\350G\\037:\\034\\313\\335\\260zZM\\220\\211\\353\\355\\357 \\\"-\\010\\001\\022)\\032\\350&\\212\\321\\322\\005 U\\267\\016\\204\\032\\361+\\271\\\"\\245\\272N\\021\\377\\313%\\375C=\\306\\256\\367\\266\\005\\033r*a\\341\\375\\232\\014 \\\"/\\010\\001\\022\\010\\034\\240B\\212\\321\\322\\005 \\032! \\355!\\316\\022tW\\031c\\260\\263e2\\\"\\345i\\203\\227\\ryP^\\001[\\326\\027\\tuQ\\235\\370\\302v\\\"0\\010\\001\\022\\t\\036\\324\\205\\001\\346\\234\\323\\005 \\032! \\335B\\276M7\\3173\\3035V'\\033\\315\\244\\200\\330 \\237j\\314b\\253\\352\\246\\304^D\\325\\345\\350\\310[\\\".\\010\\001\\022* \\200\\204\\002\\350\\260\\323\\005 \\352\\300\\001*\\345\\245\\225\\253K#//\\244\\256\\253\\327t\\023\\246\\302O\\363c\\212W\\031\\032\\3467l\\214B \\\"0\\010\\001\\022\\t$\\216\\316\\004\\350\\260\\323\\005 \\032! \\3338:Q (\\210u\\332l>\\374\\222\\310d6\\205\\321\\000\\231O\\263\\367\\266n\\372\\311\\034B\\363X\\234\\\".\\010\\001\\022*&\\366\\305\\007\\350\\260\\323\\005 \\354\\373\\303\\317\\271\\\\\\367d\\252\\346D3\\205\\244\\221\\207\\363\\346\\326\\301\\202\\203/X\\321\\251\\022\\310\\277\\241\\0308 \\\"0\\010\\001\\022\\t*\\240\\276\\022\\356\\260\\323\\005 \\032! \\027\\355I\\365Cg\\243\\374VL\\350\\003\\033Rjt\\364\\231\\375\\311\\223\\005V\\225%\\262BOG\\315\\363\\034\\\"1\\010\\001\\022\\n.\\220\\203\\206\\001\\356\\260\\323\\005 \\032! I\\262\\237\\177\\320\\342\\211\\311\\273C\\314C\\350)\\356\\257Z\\266y\\377\\315\\226y\\302\\306E\\266\\021\\257q\\322\\300\\\"/\\010\\001\\022+0\\262\\336\\254\\001\\356\\260\\323\\005 \\207{;n\\317\\337}\\2446\\235\\030dz\\367\\373\\200\\021\\014\\276\\330g\\026\\330V7\\374WM\\221\\3142# \" > ops:<type:\"ics23:simple\" key:\"acc\" data:\"\\n\\377\\001\\n\\003acc\\022 m;l\\312\\275\\\"/\\357\\373X\\325!q\\303\\343\\030\\203\\\"\\362\\023\\032\\232\\260X@\\335[\\240\\033\\263\\036Z\\032\\t\\010\\001\\030\\001 \\001*\\001\\000\\\"'\\010\\001\\022\\001\\001\\032 /Yb=\\027y\\035\\3304\\206\\207\\334\\241\\366`0A\\307\\374Fo \\244\\311m\\331\\301\\\\\\202y\\235\\274\\\"'\\010\\001\\022\\001\\001\\032 R\\324\\007\\357<xv\\342\\3566\\344o\\370\\347\\230Hl\\036\\254\\337\\237\\323\\234ps07\\275\\320=\\313F\\\"'\\010\\001\\022\\001\\001\\032 A\\227u\\330\\035\\246\\221\\014\\n\\203\\300Nk\\004oyu\\3060\\240\\006\\305\\032n\\005\\207\\303>$z:\\320\\\"'\\010\\001\\022\\001\\001\\032 o6y\\231\\276\\307+\\030\\376w\\277\\006\\357\\205ya/\\262ne\\215\\030\\3439\\2032\\374\\221\\265W\\343\\274\\\"'\\010\\001\\022\\001\\001\\032 a#v\\026\\037zEh\\222\\324\\005o\\375\\324\\364B\\321\\341\\233\\027\\005nB\\354\\\\6lb\\257\\340\\344a\" > ")
+			}
+			proofOps.Ops[i] = proof
+		}
+	}
+	return proofOps
+}
+
+func (suite *BackendTestSuite) TestGetHexProofs() {
+	defaultRes := []string{""}
+	testCases := []struct {
+		name  string
+		proof *crypto.ProofOps
+		exp   []string
+	}{
+		{
+			"no proof provided",
+			mookProofs(0, false),
+			defaultRes,
+		},
+		{
+			"no proof data provided",
+			mookProofs(1, false),
+			defaultRes,
+		},
+		{
+			"valid proof provided",
+			mookProofs(1, true),
+			[]string{"0x6f70733a3c747970653a2269637332333a6961766c22206b65793a225c3030315c3137375c725c3032355c3330375c3337325c323536655c32313166485c333130273b6d7e435c3336355c3231375c323530422220646174613a225c3032325c3331345c3032315c6e5c3032355c3030315c3137375c725c3032355c3330375c3337325c323536655c32313166485c333130273b6d7e435c3336355c3231375c323530425c3032325c3333325c3031305c6e5c3032355c3030315c3137375c725c303231585c323332575c3332365c3237345c3031345c323530666c2a775c3030335c3233306b5c32333552715c3032325c3233335c3030315c6e5c3033362f65746865726d696e742e74797065732e76312e4574684163636f756e745c303232795c6e335c6e2b7463726331307578337a6b7936326c74746372396776656b7a356163726e70346536356e337177323834665c3033305c3236305c3331325c30333620275c303232423078633564323436303138366637323333633932376537646232646363373033633065353030623635336361383232373362376266616438303435643835613437305c3033325c3031365c3031305c3030315c3033305c303031205c3030312a5c3030365c3030305c3030325c3236325c3332345c3337365c3030315c222e5c3031305c3030315c3032325c3030375c3030325c3030345c3236325c3332345c3337365c303031205c303332212029615c3235315c323236535c3236335c3231363a5c3337305c3332316c5c32343032245c3234365c3033345c3331345c3335305c3236355c303137305c33353030565c3030345c3235325c33323049545c3332346a5c3233365c222c5c3031305c3030315c303232285c3030345c3030365c3236325c3332345c3337365c303031205c3030325c32303125396f5c3233375c3237355c3230345c303335662d5c3230345c3030337c5c3234355c745c3031305c3230334d5c323333557e355c3330335c22755c3331305c3030365c32333228515c323131205c222c5c3031305c3030315c303232285c3030365c6e5c3236325c3332345c3337365c303031205c3330342d215c3332322f5c3234315c3331345c3032355c333430625c323732565c3235325c3334375c3232345c3230305c3331335c3334355c303231285c3237305c3230325c33303732755c3033345c3237314f6d5c3030355c3030335c303337205c222c5c3031305c3030315c303232285c6e5c3033345c3330305c3333365c3232325c303034205c3235305c3231375c3231365f5c3232375c3237345c3237315c3333365c3231305c3334335c3237376f5c3032355c3330315c3033372d5c3331305c323037645c3336335c3336314c5c3235335c3335375c32303735596167385c3333355c323432205c222e5c3031305c3030315c3032325c3030375c303134485c3330305c3333365c3232325c303034205c30333221205c3231343f5c3031365c323335317c5c3236355a5c3334375c3237355c3234355c323130476d5c333136315c6e5c3232305c3331365c3234333a625c3232375c3335305c3032365c3330366e5c3032305c3335366e5c3234335c3330305c222d5c3031305c3030315c303232295c3031365c3232365c3030315c3330305c3333365c3232325c303034205c333232244f5a5c3030325c323535246a3c5c3033315c3332335c3335375c3137375c3336365c333632255c323735576b5c3030315c333535405c3331325c303033265e663635495c32353167205c222f5c3031305c3030315c3032325c3031305c3032325c3333345c3030325c3331305c3233305c3332325c303035205c3033322120475c3030315c33373320524d5c32313265285c3335375c323233573b5c3230325c3237305c3337375c3333365c725c3032323b4a5c303334215c3332345c3335325c3032375c3330327c5c303232645c333636755c222f5c3031305c3030315c3032325c3031305c3032345c3330365c3030345c3331305c3233305c3332325c303035205c30333221205c3334375c3236305c303332335d5c303331435c3032335c3234335c3336315c3335375c33353527675c3230365c3232375c3235305c323631765c3334375c3336335c3334315c323032685c303231595c32303332765c333537605c3233335c222f5c3031305c3030315c3032325c3031305c3032365c3232365c3031305c3331305c3233305c3332325c303035205c30333221205c3031375c333432555c3330365a5c3234355c6e415c323632775c323035525c3332306c5c32353167773c585c323033787d5c3233363f345c3231305a5c333234645c3334325c3233365c3331305c222d5c3031305c3030315c303232295c3033305c3336345c3032345c3331305c3233305c3332325c30303520715c323033565c3232315c303333355c3335315c3332365c303230485c333531515c3336305c3334345c3335305c225c333530475c3033373a5c3033345c3331335c3333355c3236307a5a4d5c3232305c3231315c3335335c3335355c333537205c222d5c3031305c3030315c303232295c3033325c333530265c3231325c3332315c3332325c30303520555c3236375c3031365c3230345c3033325c3336312b5c3237315c225c3234355c3237324e5c3032315c3337375c333133255c333735433d5c3330365c3235365c3336375c3236365c3030355c303333722a615c3334315c3337355c3233325c303134205c222f5c3031305c3030315c3032325c3031305c3033345c323430425c3231325c3332315c3332325c303035205c30333221205c333535215c3331365c30323274575c303331635c3236305c32363365325c225c333435695c3230335c3232375c7279505e5c3030315b5c3332365c3032375c7475515c3233355c3337305c333032765c22305c3031305c3030315c3032325c745c3033365c3332345c3230355c3030315c3334365c3233345c3332335c303035205c30333221205c333335425c3237364d375c333137335c3330333556275c3033335c3331355c3234345c3230305c333330205c3233376a5c333134625c3235335c3335325c3234365c3330345e445c3332355c3334355c3335305c3331305b5c222e5c3031305c3030315c3032322a205c3230305c3230345c3030325c3335305c3236305c3332335c303035205c3335325c3330305c3030312a5c3334355c3234355c3232355c3235334b232f2f5c3234345c3235365c3235335c333237745c3032335c3234365c3330324f5c333633635c323132575c3033315c3033325c333436376c5c32313442205c22305c3031305c3030315c3032325c74245c3231365c3331365c3030345c3335305c3236305c3332335c303035205c30333221205c333333383a5120285c323130755c3333326c3e5c3337345c3232325c33313064365c3230355c3332315c3030305c3233314f5c3236335c3336375c3236366e5c3337325c3331315c303334425c333633585c3233345c222e5c3031305c3030315c3032322a265c3336365c3330355c3030375c3335305c3236305c3332335c303035205c3335345c3337335c3330335c3331375c3237315c5c5c333637645c3235325c33343644335c3230355c3234345c3232315c3230375c3336335c3334365c3332365c3330315c3230325c3230332f585c3332315c3235315c3032325c3331305c3237375c3234315c30333038205c22305c3031305c3030315c3032325c742a5c3234305c3237365c3032325c3335365c3236305c3332335c303035205c30333221205c3032375c333535495c33363543675c3234335c333734564c5c3335305c3030335c303333526a745c3336345c3233315c3337355c3331315c3232335c303035565c323235255c323632424f475c3331355c3336335c3033345c22315c3031305c3030315c3032325c6e2e5c3232305c3230335c3230365c3030315c3335365c3236305c3332335c303035205c3033322120495c3236325c3233375c3137375c3332305c3334325c3231315c3331315c323733435c333134435c333530295c3335365c3235375a5c323636795c3337375c3331355c323236795c3330325c333036455c3236365c3032315c323537715c3332325c3330305c222f5c3031305c3030315c3032322b305c3236325c3333365c3235345c3030315c3335365c3236305c3332335c303035205c3230377b3b6e5c3331375c3333377d5c323434365c3233355c303330647a5c3336375c3337335c3230305c3032315c3031345c3237365c333330675c3032365c33333056375c333734574d5c3232315c3331343223205c3033325c3332355c3031305c6e5c3032355c3030315c3137375c72705c3030305c333330507e5c33353028435c3232304d5c3235315b23625c3236325c3332315c3334315c3334315c3032325c3233315c3030315c6e5c3033362f65746865726d696e742e74797065732e76312e4574684163636f756e745c303232775c6e315c6e2b7463726331307578687171786332706c7773327a726a7078366a6b65727632656472633070396675756d6e5c3033305c3236305c3335355c3032335c303232423078633564323436303138366637323333633932376537646232646363373033633065353030623635336361383232373362376266616438303435643835613437305c3033325c725c3031305c3030315c3033305c303031205c3030312a5c3030355c3030305c3030325c3333345c3336365c3032365c222c5c3031305c3030315c303232285c3030325c3030345c3236325c3332345c3337365c30303120635c3337365c3030305c3232355c3230345c303137745c333230275c6e5c3334355c3336363a3f762d5c3237335c3230315c3231355c333135255c72565c3233302d5c3234332b5c3237375c3331355c3033365c32373228205c222c5c3031305c3030315c303232285c3030345c3030365c3236325c3332345c3337365c303031205c3030325c32303125396f5c3233375c3237355c3230345c303335662d5c3230345c3030337c5c3234355c745c3031305c3230334d5c323333557e355c3330335c22755c3331305c3030365c32333228515c323131205c222c5c3031305c3030315c303232285c3030365c6e5c3236325c3332345c3337365c303031205c3330342d215c3332322f5c3234315c3331345c3032355c333430625c323732565c3235325c3334375c3232345c3230305c3331335c3334355c303231285c3237305c3230325c33303732755c3033345c3237314f6d5c3030355c3030335c303337205c222c5c3031305c3030315c303232285c6e5c3033345c3330305c3333365c3232325c303034205c3235305c3231375c3231365f5c3232375c3237345c3237315c3333365c3231305c3334335c3237376f5c3032355c3330315c3033372d5c3331305c323037645c3336335c3336314c5c3235335c3335375c32303735596167385c3333355c323432205c222e5c3031305c3030315c3032325c3030375c303134485c3330305c3333365c3232325c303034205c30333221205c3231343f5c3031365c323335317c5c3236355a5c3334375c3237355c3234355c323130476d5c333136315c6e5c3232305c3331365c3234333a625c3232375c3335305c3032365c3330366e5c3032305c3335366e5c3234335c3330305c222d5c3031305c3030315c303232295c3031365c3232365c3030315c3330305c3333365c3232325c303034205c333232244f5a5c3030325c323535246a3c5c3033315c3332335c3335375c3137375c3336365c333632255c323735576b5c3030315c333535405c3331325c303033265e663635495c32353167205c222f5c3031305c3030315c3032325c3031305c3032325c3333345c3030325c3331305c3233305c3332325c303035205c3033322120475c3030315c33373320524d5c32313265285c3335375c323233573b5c3230325c3237305c3337375c3333365c725c3032323b4a5c303334215c3332345c3335325c3032375c3330327c5c303232645c333636755c222f5c3031305c3030315c3032325c3031305c3032345c3330365c3030345c3331305c3233305c3332325c303035205c30333221205c3334375c3236305c303332335d5c303331435c3032335c3234335c3336315c3335375c33353527675c3230365c3232375c3235305c323631765c3334375c3336335c3334315c323032685c303231595c32303332765c333537605c3233335c222f5c3031305c3030315c3032325c3031305c3032365c3232365c3031305c3331305c3233305c3332325c303035205c30333221205c3031375c333432555c3330365a5c3234355c6e415c323632775c323035525c3332306c5c32353167773c585c323033787d5c3233363f345c3231305a5c333234645c3334325c3233365c3331305c222d5c3031305c3030315c303232295c3033305c3336345c3032345c3331305c3233305c3332325c30303520715c323033565c3232315c303333355c3335315c3332365c303230485c333531515c3336305c3334345c3335305c225c333530475c3033373a5c3033345c3331335c3333355c3236307a5a4d5c3232305c3231315c3335335c3335355c333537205c222d5c3031305c3030315c303232295c3033325c333530265c3231325c3332315c3332325c30303520555c3236375c3031365c3230345c3033325c3336312b5c3237315c225c3234355c3237324e5c3032315c3337375c333133255c333735433d5c3330365c3235365c3336375c3236365c3030355c303333722a615c3334315c3337355c3233325c303134205c222f5c3031305c3030315c3032325c3031305c3033345c323430425c3231325c3332315c3332325c303035205c30333221205c333535215c3331365c30323274575c303331635c3236305c32363365325c225c333435695c3230335c3232375c7279505e5c3030315b5c3332365c3032375c7475515c3233355c3337305c333032765c22305c3031305c3030315c3032325c745c3033365c3332345c3230355c3030315c3334365c3233345c3332335c303035205c30333221205c333335425c3237364d375c333137335c3330333556275c3033335c3331355c3234345c3230305c333330205c3233376a5c333134625c3235335c3335325c3234365c3330345e445c3332355c3334355c3335305c3331305b5c222e5c3031305c3030315c3032322a205c3230305c3230345c3030325c3335305c3236305c3332335c303035205c3335325c3330305c3030312a5c3334355c3234355c3232355c3235334b232f2f5c3234345c3235365c3235335c333237745c3032335c3234365c3330324f5c333633635c323132575c3033315c3033325c333436376c5c32313442205c22305c3031305c3030315c3032325c74245c3231365c3331365c3030345c3335305c3236305c3332335c303035205c30333221205c333333383a5120285c323130755c3333326c3e5c3337345c3232325c33313064365c3230355c3332315c3030305c3233314f5c3236335c3336375c3236366e5c3337325c3331315c303334425c333633585c3233345c222e5c3031305c3030315c3032322a265c3336365c3330355c3030375c3335305c3236305c3332335c303035205c3335345c3337335c3330335c3331375c3237315c5c5c333637645c3235325c33343644335c3230355c3234345c3232315c3230375c3336335c3334365c3332365c3330315c3230325c3230332f585c3332315c3235315c3032325c3331305c3237375c3234315c30333038205c22305c3031305c3030315c3032325c742a5c3234305c3237365c3032325c3335365c3236305c3332335c303035205c30333221205c3032375c333535495c33363543675c3234335c333734564c5c3335305c3030335c303333526a745c3336345c3233315c3337355c3331315c3232335c303035565c323235255c323632424f475c3331355c3336335c3033345c22315c3031305c3030315c3032325c6e2e5c3232305c3230335c3230365c3030315c3335365c3236305c3332335c303035205c3033322120495c3236325c3233375c3137375c3332305c3334325c3231315c3331315c323733435c333134435c333530295c3335365c3235375a5c323636795c3337375c3331355c323236795c3330325c333036455c3236365c3032315c323537715c3332325c3330305c222f5c3031305c3030315c3032322b305c3236325c3333365c3235345c3030315c3335365c3236305c3332335c303035205c3230377b3b6e5c3331375c3333377d5c323434365c3233355c303330647a5c3336375c3337335c3230305c3032315c3031345c3237365c333330675c3032365c33333056375c333734574d5c3232315c33313432232022203e206f70733a3c747970653a2269637332333a73696d706c6522206b65793a226163632220646174613a225c6e5c3337375c3030315c6e5c3030336163635c303232206d3b6c5c3331325c3237355c222f5c3335375c333733585c33323521715c3330335c3334335c3033305c3230335c225c3336325c3032335c3033325c3233325c32363058405c3333355b5c3234305c3033335c3236335c3033365a5c3033325c745c3031305c3030315c3033305c303031205c3030312a5c3030315c3030305c22275c3031305c3030315c3032325c3030315c3030315c303332202f59623d5c303237795c3033355c333330345c3230365c3230375c3333345c3234315c3336366030415c3330375c333734466f205c3234345c3331316d5c3333315c3330315c5c5c323032795c3233355c3237345c22275c3031305c3030315c3032325c3030315c3030315c30333220525c3332345c3030375c3335373c78765c3334325c333536365c3334346f5c3337305c3334375c323330486c5c3033365c3235345c3333375c3233375c3332335c323334707330375c3237355c3332303d5c333133465c22275c3031305c3030315c3032325c3030315c3030315c30333220415c323237755c3333305c3033355c3234365c3232315c3031345c6e5c3230335c3330304e6b5c3030346f79755c333036305c3234305c3030365c3330355c3033326e5c3030355c3230375c3330333e247a3a5c3332305c22275c3031305c3030315c3032325c3030315c3030315c303332206f36795c3233315c3237365c3330372b5c3033305c333736775c3237375c3030365c3335375c32303579612f5c3236326e655c3231355c3033305c333433395c323033325c3337345c3232315c323635575c3334335c3237345c22275c3031305c3030315c3032325c3030315c3030315c303332206123765c3032365c3033377a45685c3232325c3332345c3030356f5c3337355c3332345c333634425c3332315c3334315c3233335c3032375c3030356e425c3335345c5c366c625c3235375c3334305c3334346122203e20"},
+		},
+	}
+	for _, tc := range testCases {
+		suite.Run(fmt.Sprintf("Case %s", tc.name), func() {
+			suite.Require().Equal(tc.exp, GetHexProofs(tc.proof))
+		})
+	}
+}
